Telephony: DataServiceStateTracker to update PROPERTY_DATA_NETWORK_TYPE setting
DataServiceStateTracker tracks the data network type and so should be the one
updating the property PROPERTY_DATA_NETWORK_TYPE in the settings database. The
change is to remove the settings update from GSM and CDMA ServiceStateTracker
and add it to DataServiceStateTracker.

Also, the constant DATA_ACCESS_LTE is mapped to thirteen instead of fourteen.

Change-Id: I331acc502283e63b2fa1fe2670f8087e72143c47